backend/requirements.txt
asgiref>=3.8.1
cachetools>=5.5.2
certifi>=2025.1.31
charset-normalizer>=3.4.1
dj-database-url>=2.3.0
dj-rest-auth>=7.0.1
Django>=5.1.6
django-allauth>=65.4.1
django-cors-headers>=4.7.0
django-filter>=25.1
djangorestframework>=3.15.2
djangorestframework_simplejwt>=5.5.0
dotenv>=0.9.9
firebase-admin>=6.7.0
google-api-core>=2.24.2
google-api-python-client>=2.164.0
google-auth>=2.38.0
google-auth-httplib2>=0.2.0
google-cloud-core>=2.4.3
google-cloud-firestore>=2.20.1
google-cloud-storage>=3.1.0
google-crc32c>=1.6.0
google-resumable-media>=2.7.2
googleapis-common-protos>=1.69.2
grpcio>=1.71.0
grpcio-status>=1.71.0
httplib2>=0.22.0
idna>=3.10
msgpack>=1.1.0
proto-plus>=1.26.1
protobuf>=5.29.3
psycopg>=3.2.5
pyasn1>=0.6.1
pyasn1_modules>=0.4.1
pycparser>=2.22
PyJWT>=2.9.0
pyparsing>=3.2.1
python-dotenv>=1.0.1
requests>=2.32.3
rsa>=4.9
sqlparse>=0.5.3
typing_extensions>=4.12.2
uritemplate>=4.1.1
urllib3>=2.3.0